+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fix a security vulnerability with how permissions were handled for some submission actions.
+- Fix sending an email notification and running an integration for a submission not requiring the “Save submissions” permission.
+- Fix an error when sending an email notification for a submission that no longer exists.
+
 ## 3.1.37 - 2026-08-29
 
 ### Fixed